move files into own class
